Expand project with loading from excel.

+++ Models/Vessel.cs
@@ -2,7 +2,7 @@
 {
     internal class Vessel
     {
-        public decimal RequiredReliability { get; set; }
+        public double RequiredReliability { get; set; }
         public LocationTime[] LocationOverTime { get; set; }
     }
 }
